+++ src/plugin-sdk/session-catalog.ts
@@ -13,13 +13,13 @@ export type {
 } from "../plugins/session-catalog.js";
 export {
   createSessionCatalogAdoptionCoordinator,
-  importSessionCatalogHistory,
   isExternalUserText,
   listAdoptedSessionCatalogSessions,
   normalizeUserText,
   sessionCatalogAdoptedSessionKey,
   sessionCatalogAdoptedSourceKey,
 } from "../plugins/session-catalog.js";
+export { importSessionCatalogHistory } from "../plugins/session-catalog-history-import.js";
 export type {
   SessionCatalog,
   SessionCatalogCapabilities,
