+++ src/plugins/session-catalog.test.ts
@@ -1,6 +1,6 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import type { OpenClawConfig } from "../config/types.openclaw.js";
-import { importSessionCatalogHistory } from "./session-catalog.js";
+import { importSessionCatalogHistory } from "./session-catalog-history-import.js";
 
 const transcript = vi.hoisted(() => ({
   messages: [] as Array<Record<string, unknown>>,
